Added nightly build script; other fixes
I added a script that uses the clone, build and report
scripts to make a low-budget nightly build. This could be
used for a CI build as well if git can be convinced to
run it somewhere when there is a push. This would be improved
if the build recorder system was modified to include time
and not just date. I also fixed the name of the gitignore file...