tweaks
Fixed an issue with -n and report where the curl commands
were send to /dev/null - it worked without -n, but was hard
to debug. Added -k to make check in build so we can see all
of the test failures, not just the first one. The nightly build
script has the OS hard coded in; not really optimal but OK.
Please enter the commit message for your changes. Lines starting