Made supressing the hyrax-dependencies repo the default
By default don't use our version of the deps so that we can
test building on CentOS using installed packages. There seems
to be an issue with fileout netcdf and having both things
present, but that's a minor point. We want to know if the code
builds using just stuff from the EPEL (and Main) yum repo.